Allow for manual session management in tests
No preinitialized will be passed to the test, so it can itself construct
multiple sessions during its run. This is needed for savegame save/load
tests. We have to improve the session cleanup; a failing test with
manual sessions can leave crucial components like Scheduler in an
unclean state.

+++ tests/game/__init__.py
@@ -20,11 +20,14 @@
 # ###################################################
 
 
+import contextlib
 import inspect
 import os
 import signal
 import tempfile
 from functools import wraps
+
+import mock
 
 # check if SIGALRM is supported, this is not the case on Windows
 # we might provide an alternative later, but for now, this will do
@@ -105,6 +108,34 @@ def create_map():
 	return savegame
 
 
+@contextlib.contextmanager
+def _dbreader_convert_dummy_objects():
+	"""
+	Wrapper around DbReader.__call__ to convert Dummy objects to valid values.
+
+	This is needed because some classes attempt to store Dummy objects in the
+	database, e.g. ConcreteObject with self._instance.getActionRuntime().
+	We fix this by replacing it with zero, SQLite doesn't care much about types,
+	and hopefully a number is less likely to break code than None.
+
+	Yes, this is ugly and will most likely break later. For now, it works.
+	"""
+	def deco(func):
+		@wraps(func)
+		def wrapper(self, command, *args):
+			args = list(args)
+			for i in range(len(args)):
+				if args[i].__class__.__name__ == 'Dummy':
+					args[i] = 0
+			return func(self, command, *args)
+		return wrapper
+
+	original = DbReader.__call__
+	DbReader.__call__ = deco(DbReader.__call__)
+	yield
+	DbReader.__call__ = original
+
+
 class SPTestSession(SPSession):
 
 	def __init__(self, db, rng_seed=None):
@@ -144,6 +175,18 @@ class SPTestSession(SPSession):
 
 		GAME_SPEED.TICKS_PER_SECOND = 16
 
+	def save(self, *args, **kwargs):
+		"""
+		Wrapper around original save function to fix some things.
+		"""
+		# SavegameManager.write_metadata tries to create a screenshot and breaks when
+		# accessing fife properties
+		with mock.patch('horizons.spsession.SavegameManager'):
+			# We need to covert Dummy() objects to a sensible value that can be stored
+			# in the database
+			with _dbreader_convert_dummy_objects():
+				return super(SPTestSession, self).save(*args, **kwargs)
+
 	def load(self, savegame, players):
 		"""
 		Stripped version of the original code. We don't need to load selections,
@@ -152,29 +195,36 @@ class SPTestSession(SPSession):
 		self.savegame = savegame
 		self.savegame_db = SavegameAccessor(self.savegame)
 
+		self.savecounter = 1
+
 		self.world = World(self)
 		self.world._init(self.savegame_db)
 		for i in sorted(players):
 			self.world.setup_player(i['id'], i['name'], i['color'], i['local'], i['is_ai'], i['difficulty'])
 		self.manager.load(self.savegame_db)
 
-	def end(self):
+	def end(self, keep_map=False):
 		"""
 		Clean up temporary files.
 		"""
 		super(SPTestSession, self).end()
 		# Find all islands in the map first
-		random_map = False
-		for (island_file, ) in self.savegame_db('SELECT file FROM island'):
-			if island_file[:7] != 'random:': # random islands don't exist as files
-				os.remove(island_file)
-			else:
-				random_map = True
-				break
+		if not keep_map:
+			for (island_file, ) in self.savegame_db('SELECT file FROM island'):
+				if island_file[:7] != 'random:': # random islands don't exist as files
+					os.remove(island_file)
 		# Finally remove savegame
 		self.savegame_db.close()
-		if not random_map:
-			os.remove(self.savegame)
+		os.remove(self.savegame)
+
+	@classmethod
+	def cleanup(cls):
+		"""
+		If a test uses manual session management, we cannot be sure that session.end was
+		called before a crash, leaving the game in an unclean state. This method should
+		return the game to a valid state.
+		"""
+		Scheduler.destroy_instance()
 
 	def run(self, ticks=1, seconds=None):
 		"""
@@ -272,6 +322,7 @@ def game_test(*args, **kwargs):
 	mapgen = kwargs.get('mapgen', create_map)
 	human_player = kwargs.get('human_player', True)
 	ai_players = kwargs.get('ai_players', 0)
+	manual_session = kwargs.get('manual_session', False)
 
 	if TEST_TIMELIMIT and timeout:
 		def handler(signum, frame):
@@ -282,13 +333,21 @@ def game_test(*args, **kwargs):
 		@wraps(func)
 		def wrapped(*args):
 			horizons.main.db = db
-			s, p = new_session(mapgen = mapgen, human_player = human_player, ai_players = ai_players)
+			if not manual_session:
+				s, p = new_session(mapgen = mapgen, human_player = human_player, ai_players = ai_players)
 			if TEST_TIMELIMIT and timeout:
 				signal.alarm(timeout)
 			try:
-				return func(s, p, *args)
+				if not manual_session:
+					return func(s, p, *args)
+				else:
+					return func(*args)
 			finally:
-				s.end()
+				if not manual_session:
+					s.end()
+				else:
+					SPTestSession.cleanup()
+
 				if TEST_TIMELIMIT and timeout:
 					signal.alarm(0)
 		return wrapped
